Changes so gcc 12.2.0 build compiles on Macintosh. 1. Don't set --target parameter on Macintosh (clang wants that, not gcc) 2. Explicitly match "clang" to identify clang on Macintosh,    because "g++" (ha ha ha) matches "clang++". 3. Don't try to default GCC_DIR on Macintosh (as that's a work around    for OpenMP detection being a bit busted for GCC builds on Windows!) 4. Replace if (1) x; else 0 with { x; } in a few macros. e.g. PROBLEM    in decenttree.cpp, and replace 0 with {}, e.g. FNJ_TRACE in    fancyrapidnj.h. 5. findPartnerForOneCluster() returns bool (and uses intptr_t rather than    int and intptr_t) to avoid unused variable and mismatching for-loop    type warnings out of the gnu compiler. 6. Likewise in Sequences::loadAlignment()     and SequenceLoader::setUpSerializedData()    (more intptr_t, less int and size_t) 7. Likewise in hammingDistance() 8. Less optimistic checking for "did we find Python3"    in pydecenttree's CMakeLists.txt 9. Returning -65536 rather than 0xFFFFFF0 in inflateMark()     in inflate.c (in zlib-1.2.7).  I'd already messed with this     (and got it wrong!) before.
Still trouble building pydecenttree, on my Macintosh laptop though.
Still having trouble finding python.

+++ sequence.cpp
@@ -402,7 +402,7 @@ bool Sequences::loadAlignment(const std::string& fastaFilePath,
     }
     std::vector<size_t> sequence_odd_site_count;
     {
-        size_t seqLen   = front().sequenceLength();
+        intptr_t seqLen   = front().sequenceLength();
         std::vector<SiteInfo> sites;
         sites.resize(seqLen);
         SiteInfo* siteData = sites.data();
@@ -410,7 +410,7 @@ bool Sequences::loadAlignment(const std::string& fastaFilePath,
         size_t seqCount = size();
         for (size_t s=0; s<seqCount; ++s) {
             const char* sequence = at(s).data();
-            for (size_t i=0; i<seqLen; ++i) {
+            for (intptr_t i=0; i<seqLen; ++i) {
                 siteData[i].handle(unknown_char, s, sequence[i]);
             }
         }
@@ -420,7 +420,7 @@ bool Sequences::loadAlignment(const std::string& fastaFilePath,
         #ifdef _OPENMP
         #pragma omp parallel for
         #endif
-        for (int i=0; i<seqLen; ++i) {
+        for (intptr_t i=0; i<seqLen; ++i) {
             SiteInfo* info = siteData + i;
             if (info->unknownCount==seqCount) {
                 continue;
@@ -470,7 +470,7 @@ void SequenceLoader::setUpSerializedData() {
         const char* read_site = sequence_data[row];
         uint64_t*   write_unk = unknown_data[row];
         uint64_t          unk = 0;
-        for (int col=0; col<seqLen; ++col) {
+        for (intptr_t col=0; col<seqLen; ++col) {
             unk <<= 1;
             if (read_site[col] == unknown_char ) {
                 ++unk;
